Add support for filter blocks
Revamp the filter system to allow for sub-clauses that are ANDed (or ORed)
together.

As part of this revamp we convert the filter display from a table to a list
and encapsulate portions of this view into AngularJS directives so as to improve
code reuse.

Some care was also taken to reduce duplication between filterdataFunctions
and the filterdata API endpoint. This required renaming filter 'keys' to
'fields' in our AngularJS views.

We also remove the distinction between 'SubProject' and 'SubProjects' filter.
The 'SubProjects' filter is now available on index.php whether or not you are
viewing the children of a single build.